Add remember me option and persistent cookies

+++ app/helpers/sessions_helper.rb
@@ -4,11 +4,23 @@ module SessionsHelper
         session[:user_id] = user.id
     end
 
+    # Remember a user in a persistant session
+    def remember(user)
+        user.remember
+        # save user_id and remember token into the cookie
+        cookies.permanent.signed[:user_id] = user.id
+        cookies.permanent[:remember_token] = user.remember_token
+    end
+
     def current_user
-        if @current_user.nil?
-            @current_user = User.find_by(id: session[:user_id])
-        else
-            @current_user
+        if (user_id = session[:user_id])
+            @current_user  = @current_user || User.find_by(id: user_id)
+        elsif (user_id = cookies.signed[:user_id])
+            user = User.find_by(id: user_id)
+            if user && user.authenticated?(cookies[:remember_token])
+                log_in user
+                @current_user = user
+            end
         end
     end
 
@@ -16,9 +28,15 @@ module SessionsHelper
         !current_user.nil?
     end
 
-    def log_out
-        session.delete(:user_id)
-        @current_user = nil
+    def forget(user)
+        user.forget
+        cookies.delete(:user_id)
+        cookies.delete(:remember_token)
     end
 
+    def log_out
+        forget(current_user)
+        session.delete(:user_id) #put this behind to preserve @current_user to use in forget or else it will goes to the elsif and login again
+        @current_user = nil
+    end
 end
